+++ app/src/main/java/com/example/hackathon/repository/AnnualGDPEntity.java
@@ -1,6 +1,5 @@
 package com.example.hackathon.repository;
 
-import androidx.annotation.NonNull;
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
@@ -19,8 +18,6 @@ public class AnnualGDPEntity {
 
     @ColumnInfo(name="usa")
     private Integer usaGDP;
-
-
 
     public Integer getYear() {
         return year;
